Move Dom listener helper to ui/
Part of #1518
